app/requirements.txt
fastapi
uvicorn
langchain_community
pydantic
langchain
pgvector
psycopg2-binary
python-dotenv
langchain_postgres
